整理七牛

+++ controllers/allArticle.go
@@ -2,7 +2,7 @@
  * @Author: QTTQ
  * @Date: 2018-10-23 11:19:50
  * @LastEditors: QTTQ
- * @LastEditTime: 2018-11-28 20:15:31
+ * @LastEditTime: 2018-11-29 15:53:25
  * @Email: 1321510155@qq.com
  */
 
@@ -19,9 +19,69 @@ import (
 	"os"
 	"paopao/models"
 	"strconv"
+	"time"
 )
 
 const DST = "static/img/"
+
+var (
+	downLoadDirPath = config1.Get("qiniu_config > url").(string)
+	config1         = parseConfig.New("config/conf.json")
+	accessKey       = config1.Get("qiniu_config > accessKey").(string)
+	secretKey       = config1.Get("qiniu_config > secretKey").(string)
+	bucket          = config1.Get("qiniu_config > bucket").(string)
+	videoDomain     = config1.Get("qiniu_config > url").(string)
+	mac             = qbox.NewMac(accessKey, secretKey)
+)
+
+func uploadQiNiu(localFile string) {
+	//取出配置文件
+	//七牛云配置
+	// downLoadDirPath := "./"
+	// accessKey := config.Get("qiniu_config > accessKey").(string)
+	// secretKey := config.Get("qiniu_config > secretKey").(string)
+	// bucket := config.Get("qiniu_config > bucket").(string)
+	fmt.Println(accessKey, secretKey, bucket, "-----mmmmm-----")
+	// videoDomain := config.Get("qiniu_config > url").(string)
+
+	putPolicy := storage.PutPolicy{
+		Scope: bucket,
+	}
+	upToken := putPolicy.UploadToken(mac)
+	fmt.Println(upToken)
+	cfg := storage.Config{}
+	//华北机房
+	cfg.Zone = &storage.ZoneHuadong
+	// 是否使用https域名
+	cfg.UseHTTPS = false
+	// 上传是否使用CDN上传加速
+	cfg.UseCdnDomains = false
+	// key := "github-x.png"
+	key := localFile
+	// 构建表单上传的对象
+	formUploader := storage.NewFormUploader(&cfg)
+	ret := storage.PutRet{}
+	// 可选配置
+	putExtra := storage.PutExtra{
+		Params: map[string]string{
+			"x:name": "github logo",
+		},
+	}
+	err := formUploader.PutFile(context.Background(), &ret, upToken, key, localFile, &putExtra)
+	if err != nil {
+		fmt.Println(err)
+		return
+	}
+	fmt.Println(ret.Key, ret.Hash)
+}
+
+func downloadQiNiu(localFile string) string {
+	domain := videoDomain
+	key := localFile
+	deadline := time.Now().Add(time.Second * 3600).Unix() //1小时有效期
+	privateAccessURL := storage.MakePrivateURL(mac, domain, key, deadline)
+	return privateAccessURL
+}
 
 func AllArticle(c *gin.Context) {
 	// page:=c.PostForm("page")
@@ -83,73 +143,35 @@ type CreatArticleParams struct {
 	Context string `form:"context" json:"context"` //文章内容
 }
 
-func uploadQiNiu(localFile string) {
-	//取出配置文件
-	config := parseConfig.New("config/conf.json")
-	//七牛云配置
-	// downLoadDirPath := "./"
-	accessKey := config.Get("qiniu_config > accessKey").(string)
-	secretKey := config.Get("qiniu_config > secretKey").(string)
-	bucket := config.Get("qiniu_config > bucket").(string)
-	fmt.Println(accessKey, secretKey, bucket, "-----mmmmm-----")
-	// videoDomain := config.Get("qiniu_config > url").(string)
-
-	mac := qbox.NewMac(accessKey, secretKey)
-	putPolicy := storage.PutPolicy{
-		Scope: bucket,
-	}
-	upToken := putPolicy.UploadToken(mac)
-	fmt.Println(upToken)
-	cfg := storage.Config{}
-	//华北机房
-	cfg.Zone = &storage.ZoneHuadong
-	// 是否使用https域名
-	cfg.UseHTTPS = false
-	// 上传是否使用CDN上传加速
-	cfg.UseCdnDomains = false
-	// key := "github-x.png"
-	key :=localFile
-	// 构建表单上传的对象
-	formUploader := storage.NewFormUploader(&cfg)
-	ret := storage.PutRet{}
-	// 可选配置
-	putExtra := storage.PutExtra{
-		Params: map[string]string{
-			"x:name": "github logo",
-		},
-	}
-	err := formUploader.PutFile(context.Background(), &ret, upToken, key, localFile, &putExtra)
-	if err != nil {
-		fmt.Println(err)
-		return
-	}
-	fmt.Println(ret.Key, ret.Hash)
-}
-
 func CreatArticle(c *gin.Context) {
 	// ua := c.GetHeader("User-Agent")
 	// ct := c.GetHeader("Content-Type")
 	// fmt.Println(ua, "-ua-", "\n", ct, "--ct--")
-
 	createParams := CreatArticleParams{}
 	c.Bind(&createParams)
 	uidIr, _ := c.Get("uid")
 	uid := uidIr.(int)
 	paths := ""
 	form, _ := c.MultipartForm()
-	files := form.File["file"]
-	fmt.Println(form, files, "--------------------")
-	for _, file := range files {
-		paths += DST + file.Filename + ","
-		// Upload the file to specific dst.
-		c.SaveUploadedFile(file, DST+file.Filename)
-		uploadQiNiu(DST + file.Filename) //上传七牛
-		err := os.Remove(DST + file.Filename)
-		if err != nil {
-			fmt.Println(err)
+	if form != nil {
+		files := form.File["file"]
+		fmt.Println(form, files, "--------------------")
+		for _, file := range files {
+			paths += downLoadDirPath + DST + file.Filename + ","
+			// Upload the file to specific dst.
+			c.SaveUploadedFile(file, DST+file.Filename)
+			uploadQiNiu(DST + file.Filename) //上传七牛
+			err := os.Remove(DST + file.Filename)
+			if err != nil {
+				fmt.Println(err)
+			}
 		}
 	}
 
+	if len(paths) > 0 {
+		rs := []rune(paths)
+		paths = string(rs[0 : len(rs)-1])
+	}
 	article, err := models.CreatArticle(uid, createParams.Title, createParams.Context, paths)
 	if err != nil {
 		c.JSON(http.StatusOK, ApiRes{
@@ -167,49 +189,6 @@ func CreatArticle(c *gin.Context) {
 	})
 	return
 }
-
-// type CreatArticleParams struct {
-// 	// Uid     string `form:"uid" json:"uid"`         //uid
-// 	Title   string   `form:"title" json:"title"`     //文章主题
-// 	Context string   `form:"context" json:"context"` //文章内容
-// 	Paths   []string `form:"paths" json:"paths"`     //地址
-
-// }
-
-// func CreatArticle(c *gin.Context) {
-// 	fmt.Println("===============111111111111====================")
-
-// 	createParams := CreatArticleParams{}
-// 	c.Bind(&createParams)
-// 	paths := ""
-// 	fmt.Println(createParams.Context, "===================================")
-// 	uidIr, _ := c.Get("uid")
-
-// 	uid := uidIr.(int)
-// 	c.JSON(http.StatusOK, ApiRes{
-// 		Code: 1,
-// 		Msg:  "获取uid失败",
-// 	})
-// 	for _, file := range createParams.Paths {
-// 		paths += file + ","
-// 	}
-// 	article, err := models.CreatArticle(uid, createParams.Title, createParams.Context, paths)
-// 	if err != nil {
-// 		c.JSON(http.StatusOK, ApiRes{
-// 			Code: 1,
-// 			Msg:  "登录失败",
-// 		})
-// 		return
-// 	}
-// 	c.JSON(http.StatusOK, ApiRes{
-// 		Code: 0,
-// 		Msg:  "创建文章成功",
-// 		Data: gin.H{
-// 			"data": article,
-// 		},
-// 	})
-// 	return
-// }
 
 func DeleteArticle(c *gin.Context) {
 	deleteParams := MyArticleParams{}
